+++ sdk/communication/azure-communication-callingserver/src/main/java/com/azure/communication/callingserver/CallingServerServiceVersion.java
@@ -6,11 +6,10 @@ package com.azure.communication.callingserver;
 import com.azure.core.util.ServiceVersion;
 
 /**
- * The versions of CallingServer Service supported by this client library.
+ * Versions of CallingServer service supported by this client library.
  */
 public enum CallingServerServiceVersion implements ServiceVersion {
-    V2021_06_015("2021-06-15"),
-    V2021_06_15_preview("2021-06-15-preview");
+    V2021_06_15_PREVIEW("2021-06-15-preview");
 
     private final String version;
 
@@ -29,9 +28,9 @@ public enum CallingServerServiceVersion implements ServiceVersion {
     /**
      * Gets the latest service version supported by this client library
      *
-     * @return the latest {@link CallingServerServiceVersion}
+     * @return The latest {@link CallingServerServiceVersion} object.
      */
     public static CallingServerServiceVersion getLatest() {
-        return V2021_06_15_preview;
+        return V2021_06_15_PREVIEW;
     }
 }