+++ sdk/keyvault/azure-security-keyvault-jca/src/test/java/com/azure/security/keyvault/jca/KeyVaultCertificatesTest.java
@@ -18,21 +18,20 @@ public class KeyVaultCertificatesTest {
 
     private final KeyVaultClient keyVaultClient = mock(KeyVaultClient.class);
 
-    private Key key = mock(Key.class);
+    private final Key key = mock(Key.class);
 
-    private Certificate certificate = mock(Certificate.class);
+    private final Certificate certificate = mock(Certificate.class);
 
     private KeyVaultCertificates keyVaultCertificates;
 
     @BeforeEach
     public void beforeEach() {
-
         List<String> aliases = new ArrayList<>();
         aliases.add("myalias");
         when(keyVaultClient.getAliases()).thenReturn(aliases);
         when(keyVaultClient.getKey("myalias", null)).thenReturn(key);
         when(keyVaultClient.getCertificate("myalias")).thenReturn(certificate);
-        keyVaultCertificates = new KeyVaultCertificates(0, keyVaultClient);
+        keyVaultCertificates = new KeyVaultCertificates(60_000, keyVaultClient);
     }
 
     @Test
@@ -53,8 +52,10 @@ public class KeyVaultCertificatesTest {
     @Test
     public void testRefreshAndGetAliasByCertificate() {
         Assertions.assertEquals(keyVaultCertificates.refreshAndGetAliasByCertificate(certificate), "myalias");
+        Assertions.assertEquals(keyVaultCertificates.getCertificates().get("myalias"), certificate);
         when(keyVaultClient.getAliases()).thenReturn(null);
         Assertions.assertNotEquals(keyVaultCertificates.refreshAndGetAliasByCertificate(certificate), "myalias");
+        Assertions.assertNull(keyVaultCertificates.getCertificates().get("myalias"));
     }
 
     @Test
@@ -62,21 +63,6 @@ public class KeyVaultCertificatesTest {
         Assertions.assertTrue(keyVaultCertificates.getAliases().contains("myalias"));
         keyVaultCertificates.deleteEntry("myalias");
         Assertions.assertFalse(keyVaultCertificates.getAliases().contains("myalias"));
-    }
-
-    @Test
-    public void testCertificatesNeedRefresh() throws InterruptedException {
-        keyVaultCertificates = new KeyVaultCertificates(1000, keyVaultClient);
-        Assertions.assertTrue(keyVaultCertificates.certificatesNeedRefresh());
-        keyVaultCertificates.getAliases();
-        Assertions.assertFalse(keyVaultCertificates.certificatesNeedRefresh());
-        Thread.sleep(10);
-        KeyVaultCertificates.updateLastForceRefreshTime();
-        Assertions.assertTrue(keyVaultCertificates.certificatesNeedRefresh());
-        keyVaultCertificates.getAliases();
-        Assertions.assertFalse(keyVaultCertificates.certificatesNeedRefresh());
-        Thread.sleep(2000);
-        Assertions.assertTrue(keyVaultCertificates.certificatesNeedRefresh());
     }
 
 }