+++ sdk/spring/azure-spring-cloud-context/src/main/java/com/azure/spring/cloud/autoconfigure/context/AzureContextAutoConfiguration.java
@@ -8,9 +8,10 @@ import com.azure.core.management.profile.AzureProfile;
 import com.azure.resourcemanager.AzureResourceManager;
 import com.azure.spring.cloud.context.core.api.CredentialsProvider;
 import com.azure.spring.cloud.context.core.api.EnvironmentProvider;
-import com.azure.spring.core.AzureSpringProperties;
 import com.azure.spring.core.CredentialProperties;
+import com.azure.spring.cloud.context.core.impl.ResourceGroupManager;
 import com.azure.spring.identity.DefaultSpringCredentialBuilder;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
@@ -63,5 +64,18 @@ public class AzureContextAutoConfiguration {
         return new DefaultSpringCredentialBuilder().environment(environment)
                                                    .alternativePrefix(AzureContextProperties.PREFIX)
                                                    .build();
+    }
+
+    @Bean
+    @ConditionalOnMissingBean
+    @ConditionalOnBean(AzureResourceManager.class)
+    public ResourceGroupManager resourceGroupManager(AzureResourceManager azureResourceManager,
+                                                         AzureContextProperties azureContextProperties) {
+        ResourceGroupManager resourceGroupManager = new ResourceGroupManager(azureResourceManager, azureContextProperties);
+        if (azureContextProperties.isAutoCreateResources()
+            && !resourceGroupManager.exists(azureContextProperties.getResourceGroup())) {
+            resourceGroupManager.create(azureContextProperties.getResourceGroup());
+        }
+        return resourceGroupManager;
     }
 }