+++ sdk/tables/azure-data-tables/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 12.1.0-beta.1 (Unreleased)
 
+### Bug fixes
+- Fixed issue where HTTP headers set in a `ClientOptions` object passed to a client builder would not be set on a client instantiated by said builder.
+- Fixed an issue where a `connectionString` with an account name and key would override a `sasToken`'s authentication settings in client builders.
+- Fixed an issue that made `TableClient.listEntities()` and `TableServiceClient.listTables()` throw a `ClassCastException` when passing a non-null value for `timeout`.
 
 ## 12.0.0 (2021-06-11)